requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
lxml
gmplot
scikit-learn scikit-learn>=0.19.0
numpy>=1.13.1
pandas>=0.20
seaborn>=0.8
matplotlib>=2.0.2
